+++ src/model_entity_color_renderer.cpp
@@ -100,7 +100,7 @@ void ModelEntityColorRenderer::processPointlightEntities(const unordered_map<str
 		}
 	}
 
-	for(size_t i = 0; i < visibleEntities.size(); i++)
+	for(unsigned int i = 0; i < visibleEntities.size(); i++)
 	{
 		_shader->uploadUniform("u_pointlightPositions[" + to_string(i) + "]", visibleEntities[i]->getPosition());
 		_shader->uploadUniform("u_pointlightColors[" + to_string(i) + "]", visibleEntities[i]->getColor());
@@ -123,7 +123,7 @@ void ModelEntityColorRenderer::processSpotlightEntities(const unordered_map<stri
 		}
 	}
 
-	for(size_t i = 0; i < visibleEntities.size(); i++)
+	for(unsigned int i = 0; i < visibleEntities.size(); i++)
 	{
 		_shader->uploadUniform("u_spotlightPositions[" + to_string(i) + "]", visibleEntities[i]->getPosition());
 		_shader->uploadUniform("u_spotlightFronts[" + to_string(i) + "]", visibleEntities[i]->getFront());
