+++ src/water_entity_color_renderer.cpp
@@ -102,7 +102,7 @@ void WaterEntityColorRenderer::processPointlightEntities(const unordered_map<str
 	}
 
 	unsigned int index = 0;
-	for(size_t i = 0; i < visibleEntities.size(); i++)
+	for(unsigned int i = 0; i < visibleEntities.size(); i++)
 	{
 		//_shader->uploadUniform("u_pointlightPositions[" + to_string(i) + "]", visibleEntities[i]->getPosition());
 		//_shader->uploadUniform("u_pointlightColors[" + to_string(i) + "]", visibleEntities[i]->getColor());
@@ -125,7 +125,7 @@ void WaterEntityColorRenderer::processSpotlightEntities(const unordered_map<stri
 		}
 	}
 
-	for(size_t i = 0; i < visibleEntities.size(); i++)
+	for(unsigned int i = 0; i < visibleEntities.size(); i++)
 	{
 		//_shader->uploadUniform("u_spotlightPositions[" + to_string(i) + "]", visibleEntities[i]->getPosition());
 		//_shader->uploadUniform("u_spotlightFronts[" + to_string(i) + "]", visibleEntities[i]->getFront());
